add tree search, add example, fix spacing
