+++ README.md
@@ -17,7 +17,7 @@ There are several different sub-modules in anyrl:
  * `models`: abstractions and concrete implementations of RL models. This includes actor-critic RNNs, MLPs, CNNs, etc. Takes care of sequence padding, BPTT, etc.
  * `envs`: APIs for dealing with environments, including wrappers and asynchronous environments.
  * `rollouts`: APIs for gathering and manipulating batches of episodes or partial episodes. Many RL algorithms include a "gather trajectories" step, and this sub-module fulfills that role.
- * `algos`: well-known learning algorithms like policy gradients or PPO. Also includes mini-algorithms like Generalized Advantage Estimation.
+ * `algos`: well-known learning algorithms like policy gradients or PPO. Also includes mini-algorithms like Generalized Advantage Estimation and Monte Carlo Tree Search.
  * `spaces`: tools for using action and observation spaces. Includes parameterized probability distributions for implementing stochastic policies.
 
 # Motivation
