+++ anyrl/algos/__init__.py
@@ -6,6 +6,7 @@ from .advantages import AdvantageEstimator, GAE
 from .a2c import A2C
 from .ppo import PPO
 from .schedules import TFSchedule, LinearTFSchedule
+from .tree import TreeNode, TreeSearch, MonteCarloTreeSearch
 
 # Don't import mpi by default since that breaks when
 # you don't run your program with `mpirun`.
