+++ anyrl/rollouts/__init__.py
@@ -5,7 +5,7 @@ Various ways to gather trajectories in RL environments.
 from .list import mean_total_reward, mean_finished_reward
 from .logger import EpisodeLogger
 from .norm import RewardNormalizer
-from .rollers import Roller, BasicRoller, TruncatedRoller, EpisodeRoller
+from .rollers import Roller, BasicRoller, TruncatedRoller, EpisodeRoller, TreeRoller
 from .rollout import Rollout, empty_rollout
 
 __all__ = dir()
