--- anyrl/rollouts/search/__init__.py
@@ -1,6 +1,0 @@
-"""
-A set of APIs for guided action space search
-"""
-
-
-__all__ = dir()
